% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/createBaiduLBS.R
\name{createBaiduLBS}
\alias{createBaiduLBS}
\title{Create an Baidu LBS object}
\usage{
createBaiduLBS(app_id)
}
\arguments{
\item{app_id}{ID of the application.}
}
\value{
An reference object of \code{\link{BaiduLBS}}.
}
\description{
Create an authorized object
}
\note{
There is only one BaiduLBS object needed.
}
\examples{
\dontrun{

baidu <- createBaiduLBS("1234567")
}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\references{
\url{http://developer.baidu.com/map}
}
\seealso{
\code{\link{registerKey}}
}
\keyword{authorization}

