% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getGeocoding.R
\name{Geocoding API}
\alias{Geocoding API}
\alias{getGeocoding}
\alias{revGeocoding}
\title{Geocoding API.}
\usage{
getGeocoding(address, city, ak = NULL)

revGeocoding(location, coordtype = c("bd09ll", "gcj02ll", "wgs84ll"),
  around = FALSE, ak = NULL)
}
\arguments{
\item{address}{String of address.}

\item{city}{City name.}

\item{ak}{Access key of an application.}

\item{location}{A numeric vecter with 2 elements indicating latitude and longitude.}

\item{coordtype}{Coordinate type, one of the "bd09ll", "gcj02ll" and "wgs84ll".}

\item{around}{Whether to output the POIs around.}
}
\value{
a data.frame.
}
\description{
Get geocoding of an address or get the address of geocoding.
}
\examples{
\dontrun{
getGeocoding("KFC")
revGeocoding(c(22.3087, 114.2019))
}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\references{
\url{http://developer.baidu.com/map/index.php?title=webapi/guide/webservice-geocoding}
}

