% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geoconv.R
\name{geoconv}
\alias{geoconv}
\title{Convert other coordinates to Baidu coordinate.}
\usage{
geoconv(geodf, from = c("GPS", "GPSmc", "google", "googlemc", "bd09ll",
  "bd09mc", "mapbar", "51map"), to = c("bd09ll", "bd09mc"), lonCol = "Lon",
  latCol = "Lat", ak = NULL)
}
\arguments{
\item{geodf}{data frame to be converted.}

\item{from}{From what coordinate.}

\item{to}{To what coordinate.}

\item{lonCol}{Column name of the longitude.}

\item{latCol}{Column name of the latitude.}
}
\value{
a data.frame.
}
\description{
Convert other coordinates to Baidu coordinate.
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\references{
\url{http://developer.baidu.com/map/index.php?title=webapi/guide/changeposition}
}

