% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getStaticImage.R
\name{getStaticImage}
\alias{getStaticImage}
\title{Static image API.}
\usage{
getStaticImage(location, outfile = "staticimage.png", width = 400,
  height = 300, zoom = 19, labels = FALSE, labelcontent = "<U+00B3><U+0279><U+00A6><c2><ca>: ",
  labelfontsize = 14, labelfontcol = "0xffffff", labelbgcol = "0x000fff")
}
\arguments{
\item{location}{A numeric vecter with 2 elements indicating latitude and longitude.}

\item{outfile}{Path of the output png file.}

\item{width}{Width of the image.}

\item{height}{Height of the image.}

\item{zoom}{Zoom size.}
}
\value{
The path of output file.
}
\description{
Get static image and save it as png file.
}
\examples{
\dontrun{
getStaticImage(c(22.3087, 114.2019))
}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\references{
\url{http://developer.baidu.com/map/index.php?title=static}
}

