% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/manageKey.R
\name{Manage keys}
\alias{Manage keys}
\alias{deleteKey}
\alias{listKeys}
\alias{modifyKey}
\alias{registerKey}
\title{Reigster, modify and delete application.}
\usage{
registerKey(app_id, app_name, app_ak)

modifyKey(app_id, app_name, app_ak)

deleteKey(app_id)

listKeys(app_id)
}
\arguments{
\item{app_id}{ID of an application}

\item{app_name}{name of an application}

\item{app_ak}{access key of an application}
}
\value{
a logical value
}
\description{
Reigster, modify and delete application.
}
\note{
You should register an application on baidu firstly.
}
\examples{
\dontrun{
registerKey(app_id = "1234567", "test", "MCD8BKwGdgPHvAuvgvz4EQpD")
listKeys("1234567")
modifyKey("1234567", "t1", "t2")
deleteKey("1234567")
}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\references{
\url{http://developer.baidu.com/map}
}
\keyword{authorization}

