% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/searchDirection.R
\name{searchDirection}
\alias{searchDirection}
\title{Search navigation scheme.}
\usage{
searchDirection(origin, destination, region, mode = c("driving", "walking",
  "transit"), origin_region, destination_region, coordtype = c("bd09ll",
  "gcj02ll", "wgs84ll"), tactics = c("nohighway", "mintime", "mindist"),
  ak = NULL)
}
\arguments{
\item{origin}{Origin place.}

\item{destination}{Destination place.}

\item{region}{City of the places. If \code{mode} is "walking" or "transit", it's required.}

\item{mode}{Method of transport, one of "driving", "walking" and "transit".}

\item{origin_region}{City of the origin place. If \code{mode} is "driving", it's required.}

\item{destination_region}{City of the destination place. If \code{mode} is "driving", it's required.}

\item{coordtype}{Coordinate type, one of the "bd09ll", "gcj02ll" and "wgs84ll".}

\item{tactics}{Navigation tactic, "nohighway" means avoiding highway,
"mintime" means shortest time, "mindist" means shortest path.}

\item{ak}{Access key of an application.}
}
\value{
A list of the details.
}
\description{
Search navigation scheme.
}
\examples{
\dontrun{
searchDirection(origin = "RUC", destination = "PKU", mode = "walking", region = "Beijing")
}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\references{
\url{http://developer.baidu.com/map/index.php?title=webapi/direction-api}
}

