% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/searchPlace.R
\name{searchPlace}
\alias{searchPlace}
\title{Search detail of a place.}
\usage{
searchPlace(q, region, bounds, location, radius, page_num = 1,
  page_size = 20, detail = TRUE, ak = NULL)
}
\arguments{
\item{q}{Query string.}

\item{region}{City name.}

\item{bounds}{A numeric vecter with 4 elements indicating latitude of bottom left,
longitude of bottom left, latitude of upper right, longitude of upper right.}

\item{location}{A numeric vecter with 2 elements indicating latitude and longitude of the centre of a circle.}

\item{radius}{The radius in meters.}

\item{page_num}{Which page of the results should be outputed.}

\item{page_size}{How many items in each page. Default is 20.}

\item{ak}{Access key of an application.}
}
\value{
A data.frame of the details:
 \item{uid}{ID of POI.}
 \item{name}{Name of POI.}
 \item{lat}{Latitude.}
 \item{lng}{Longitude}
 \item{address}{Address.}
 \item{telephone}{Telephone number.}
 \item{distance}{Distance to the centre.}
 \item{type}{Type of POI, such as 'hotel' and 'cater'.}
 \item{tag}{Tag.}
 \item{price}{Price of the business.}
 \item{shop_hours}{Business hours.}
 \item{overall_rating}{Overall rating.}
 \item{taste_rating}{Taste rating.}
 \item{service_rating}{Service rating.}
 \item{environment_rating}{Environment rating.}
 \item{facility_rating}{Facility rating.}
 \item{hygiene_rating}{Hygiene rating.}
 \item{technology_rating}{Technology rating.}
 \item{image_num}{Image number.}
 \item{groupon_num}{Groupon number.}
 \item{discount_num}{Discount number.}
 \item{comment_num}{Comments number.}
 \item{favorite_num}{Favorite number.}
 \item{checkin_num}{Checkin number.}
 \item{detail_url}{URL of the page.}
}
\description{
Search detail of a place.
}
\details{
One of the arguments 'region', 'bounds' and 'location/radius' should be specified.
}
\examples{
\dontrun{
searchPlace("KFC")
searchPlace("KFC", bounds = c(39.915, 116.404, 39.975, 116.414))
searchPlace("KFC", location = c(39.915, 116.404), radius = 2000)
}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\references{
\url{http://developer.baidu.com/map/index.php?title=webapi/guide/webservice-placeapi}
}

