% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/searchPOI.R
\name{searchPOI}
\alias{searchPOI}
\title{Search detail of a POI.}
\usage{
searchPOI(uid, ak = NULL)
}
\arguments{
\item{uid}{ID of POI.}

\item{ak}{Access key of an application.}
}
\value{
A data.frame of the details:
 \item{uid}{ID of POI.}
 \item{name}{Name of POI.}
 \item{lat}{Latitude.}
 \item{lng}{Longitude}
 \item{address}{Address.}
 \item{telephone}{Telephone number.}
 \item{distance}{Distance to the centre.}
 \item{type}{Type of POI, such as 'hotel' and 'cater'.}
 \item{tag}{Tag.}
 \item{price}{Price of the business.}
 \item{shop_hours}{Business hours.}
 \item{overall_rating}{Overall rating.}
 \item{taste_rating}{Taste rating.}
 \item{service_rating}{Service rating.}
 \item{environment_rating}{Environment rating.}
 \item{facility_rating}{Facility rating.}
 \item{hygiene_rating}{Hygiene rating.}
 \item{technology_rating}{Technology rating.}
 \item{image_num}{Image number.}
 \item{groupon_num}{Groupon number.}
 \item{discount_num}{Discount number.}
 \item{comment_num}{Comments number.}
 \item{favorite_num}{Favorite number.}
 \item{checkin_num}{Checkin number.}
 \item{atmosphere}{Atmosphere.}
 \item{featured_service}{Featured service.}
 \item{alias}{Alias of the POI.}
 \item{description}{Description of the POI.}
 \item{detail_url}{URL of the page.}
}
\description{
Search detail of a POI.
}
\examples{
\dontrun{
searchPOI("72aa251304ef0b5e067ce896")
}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\references{
\url{http://developer.baidu.com/map/index.php?title=webapi/guide/webservice-placeapi}
}

